pipeline/pyproject.toml
[tool.poetry]
name = "pipeline"
version = "0.1.0"
description = ""
authors = ["haojunsng <snghaojun@outlook.sg>"]
readme = "README.md"
packages = [{include = "pipeline"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
requests = "^2.31.0"
boto3 = "^1.34.74"
psycopg2-binary = "^2.9.9"
apache-airflow = {extras = ["amazon"], version = "^2.9.0"}
pandas = "2.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
